change total_mark to float

+++ db/migrate/20120805131713_add_total_mark_to_response_set.rb
@@ -1,5 +1,5 @@
 class AddTotalMarkToResponseSet < ActiveRecord::Migration
   def change
-    add_column :response_sets, :total_mark, :integer
+    add_column :response_sets, :total_mark, :float
   end
 end
